refactor(app): shift group column table  WIP
